booxclash-pro/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
websockets
# The critical package for your error:
google-generativeai
# The newer Google SDK (keep if you are experimenting with it, otherwise optional):
google-genai
# Vector database:
pinecone
httpx
pdfplumber
firebase_admin
cloudinary
jinja2
python-multipart
fuzzywuzzy